# docs/swagger.yaml
basePath: /
definitions:
  buildinfo.Info:
    properties:
      build_time:
        type: string
      commit:
        type: string
      started_at:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
  health.Dependency:
    properties:
      latency:
        type: string
      status:
        type: string
    type: object
  health.Status:
    properties:
      dependencies:
        additionalProperties:
          $ref: '#/definitions/health.Dependency'
        type: object
      status:
        type: string
    type: object
  httptransport.InstanceBody:
    properties:
      endpoint:
        type: string
      instance_id:
        type: string
      lease_expires_at:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      protocol:
        type: string
      service_name:
        type: string
      status:
        type: string
      version:
        type: string
      weight:
        type: integer
    type: object
  httptransport.ListInstancesBody:
    properties:
      instances:
        items:
          $ref: '#/definitions/httptransport.InstanceBody'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      revision:
        type: integer
      total:
        type: integer
    type: object
  httptransport.ListInstancesRequest:
    properties:
      include_draining:
        type: boolean
      metadata:
        additionalProperties:
          type: string
        type: object
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
      service_name:
        type: string
    type: object
  httptransport.ListServicesBody:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      revision:
        type: integer
      services:
        items:
          $ref: '#/definitions/httptransport.ServiceSummaryBody'
        type: array
      total:
        type: integer
    type: object
  httptransport.ListServicesRequest:
    properties:
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
      prefix:
        type: string
    type: object
  httptransport.MeResponseBody:
    properties:
      subject:
        type: string
    type: object
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
  httptransport.ServiceSummaryBody:
    properties:
      draining_instances:
        type: integer
      healthy_instances:
        type: integer
      service_name:
        type: string
    type: object
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
  /api/v1/me:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.MeResponseBody'
              type: object
        "401":
          description: 'Code 20001: unauthorized'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Return the authenticated subject
      tags:
      - authentication
  /api/v1/registry/instances/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service name or metadata selector
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListInstancesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ListInstancesBody'
              type: object
        "400":
          description: 'Code 10001: invalid selector'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: List live service instances
      tags:
      - registry
  /api/v1/registry/services/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Optional service prefix
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListServicesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ListServicesBody'
              type: object
      security:
      - Bearer: []
      summary: List registered services
      tags:
      - registry
  /api/v1/version:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/buildinfo.Info'
              type: object
      summary: Return build and runtime version information
      tags:
      - operations
  /live:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
      summary: Check process liveness
      tags:
      - operations
  /ready:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
        "503":
          description: 'Code 50003: dependency unavailable'
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
      summary: Check database and Redis readiness
      tags:
      - operations
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"